+++ modules/canvas-kit-react-core/index.ts
@@ -1,7 +1,7 @@
 import * as canvasColorsWeb from '@workday/canvas-colors-web';
 import {BrandingColors} from './lib/colors.types';
 import depth, {CanvasDepth} from './lib/depth';
-import spacing, {CanvasSpacing} from './lib/spacing';
+import spacing, {spacingNumbers, CanvasSpacingNumbers, CanvasSpacing} from './lib/spacing';
 import type, {fontFamily, monoFontFamily, CanvasType} from './lib/type';
 import {CSSProperties} from './lib/types';
 
@@ -21,12 +21,14 @@ export {
   colors,
   depth,
   spacing,
+  spacingNumbers,
   type,
   fontFamily,
   monoFontFamily,
   BrandingColors,
   CanvasDepth,
   CanvasSpacing,
+  CanvasSpacingNumbers,
   CanvasType,
   CSSProperties,
 };