+++ modules/canvas-kit-react-header/lib/shared/types.tsx
@@ -1,0 +1,15 @@
+export enum HeaderTheme {
+  White,
+  Blue,
+  Transparent,
+}
+
+export enum HeaderVariant {
+  Full,
+  Dub,
+}
+
+export enum HeaderHeight {
+  Small = '64px',
+  Large = '80px',
+}