+++ typings.d.ts
@@ -3,4 +3,6 @@ declare module '*.md' {
   export default value;
 }
 
+declare module '*.jpg';
+
 declare module 'storybook-readme/with-readme';