+++ modules/canvas-kit-react-common/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.0.7"></a>
+## [0.0.7](https://ghe.megaleo.com/design/canvas-kit-react/tree/master/modules/canvas-kit-react-common/compare/@workday/canvas-kit-react-common@0.0.6...@workday/canvas-kit-react-common@0.0.7) (2018-10-24)
+
+
+
+
+**Note:** Version bump only for package @workday/canvas-kit-react-common
+
 <a name="0.0.6"></a>
 ## [0.0.6](https://ghe.megaleo.com/design/canvas-kit-react/tree/master/modules/canvas-kit-react-common/compare/@workday/canvas-kit-react-common@0.0.5...@workday/canvas-kit-react-common@0.0.6) (2018-10-19)